2002-06-08  Roland McGrath  <roland@frob.com>
	* sysdeps/mach/hurd/dirstream.h (struct __dirstream): Use size_t
	instead of unsigned long int for __allocation and __size members.

+++ sysdeps/mach/hurd/dirstream.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 1993, 1994, 1995, 1996, 1997 Free Software Foundation, Inc.
+/* Copyright (C) 1993,94,95,96,97,2002 Free Software Foundation, Inc.
    This file is part of the GNU C Library.
 
    The GNU C Library is free software; you can redistribute it and/or
@@ -34,8 +34,8 @@ struct __dirstream
     int __entry_data;		/* Entry number `__data' corresponds to.  */
     char *__ptr;		/* Current pointer into the block.  */
     int __entry_ptr;		/* Entry number `__ptr' corresponds to.  */
-    unsigned long int __allocation; /* Space allocated for the block.  */
-    unsigned long int __size;	/* Total valid data in the block.  */
+    size_t __allocation;	/* Space allocated for the block.  */
+    size_t __size;		/* Total valid data in the block.  */
     __libc_lock_define (, __lock) /* Mutex lock for this structure.  */
   };
